user dashboard api

+++ app/Http/Controllers/Api/AuthController.php
@@ -3,12 +3,15 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Api\BaseController;
+use App\Http\Resources\DriverProfile;
 use App\Http\Resources\UserProfileCollection;
+use App\Models\Setting;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 // use App\Http\Resources\UserProfileCollection;
 use App\Models\User;
 use Auth;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Storage;
 use Validator;
 
@@ -18,6 +21,46 @@ class AuthController extends BaseController
 	public function unauthorized_access()
 	{
 		return $this->sendFailed('YOU ARE NOT UNAUTHORIZED TO ACCESS THIS URL, PLEASE LOGIN AGAIN', 401);
+	}
+
+	public function registerOtpVerify(Request $request)
+	{
+		$error_message = 	[
+			// 'mobile.unique'  	              => 'mobile has been already taken',
+			'mobile.required'            	  => 'Mobile should be required',
+			'user_id.required'			 	  => 'User Id should be required',
+			'otp.required'					  => 'OTP should be required',
+		];
+		$rules = [
+			'mobile'                          => 'required|min:10|max:10',
+			'user_id'						  => 'required|integer|exists:users,id',
+			'otp'						      => 'required|integer'
+		];
+		$validator = Validator::make($request->all(), $rules, $error_message);
+		if ($validator->fails()) {
+			return $this->sendFailed($validator->errors()->first(), 200);
+		}
+		try {
+			\DB::beginTransaction();
+
+			$user_detail = User::find($request->user_id);
+			if ($user_detail->mobile != $request->mobile) {
+				return $this->sendFailed("Mobile number does not exist", 200);
+			}
+			if ($request->otp != $user_detail->otp) {
+				return $this->sendFailed("wrong otp", 200);
+			}
+			// $user = User::updateOrCreate(
+			// 	$request->only('mobile', 'role')
+			// );
+			Auth::loginUsingId($user_detail->id);
+			$access_token = auth()->user()->createToken(auth()->user()->mobile)->accessToken;
+			\DB::commit();
+			return $this->sendSuccess('LOGGED IN SUCCESSFULLY', ['access_token' => $access_token, 'profile_data' => new UserProfileCollection(auth()->user())]);
+		} catch (\Throwable $e) {
+			\DB::rollback();
+			return $this->sendFailed($e->getMessage() . ' on line ' . $e->getLine(), 400);
+		}
 	}
 
 	// CREATE ACCOUNT API
@@ -34,7 +77,6 @@ class AuthController extends BaseController
 			'role'						      => 'required|In:2,3'
 		];
 		$validator = Validator::make($request->all(), $rules, $error_message);
-
 		if ($validator->fails()) {
 			return $this->sendFailed($validator->errors()->first(), 200);
 		}
@@ -53,10 +95,16 @@ class AuthController extends BaseController
 			$user = User::updateOrCreate(
 				$request->only('mobile', 'role')
 			);
-			Auth::loginUsingId($user->id);
-			$access_token = auth()->user()->createToken(auth()->user()->mobile)->accessToken;
+			// print_r($user->id);die;
+			$verifaction_otp = rand(1000, 9999);
+			$user->otp = $verifaction_otp;
+			if ($user->role == 3 && $user->unique_id == '') {
+				$user->unique_id = $user->id . rand(1000, 9999);
+			}
+			$user->save();
+			Self::send_sms_otp($request->mobile, $verifaction_otp);
 			\DB::commit();
-			return $this->sendSuccess('LOGGED IN SUCCESSFULLY', ['access_token' => $access_token, 'profile_data' => new UserProfileCollection(auth()->user())]);
+			return $this->sendSuccess('OTP SENT SUCCESSFULLY', ['user_id' => $user->id, 'otp' => $verifaction_otp]);
 		} catch (\Throwable $e) {
 			\DB::rollback();
 			return $this->sendFailed($e->getMessage() . ' on line ' . $e->getLine(), 400);
@@ -66,6 +114,12 @@ class AuthController extends BaseController
 	public function getUserProfile()
 	{
 		return $this->sendSuccess('LOGGED IN SUCCESSFULLY', ['profile_data' => new UserProfileCollection(auth()->user())]);
+	}
+
+	public function getDriverProfile()
+	{
+		$delivery_charge                = Setting::value('deliver_charge');
+		return $this->sendSuccess('LOGGED IN SUCCESSFULLY', ['profile_data' => new DriverProfile(auth()->user())]);
 	}
 
 	public function sentRegisterOtp(Request $request)
@@ -81,9 +135,8 @@ class AuthController extends BaseController
 			return $this->sendFailed($validator->errors()->first(), 200);
 		}
 		try {
-
-			$verifaction_otp = rand(111111, 999999);
-			$email_data = ['verifaction_otp' => $verifaction_otp];
+			$verifaction_otp = rand(1000, 9999);
+			$email_data = ['otp' => $verifaction_otp];
 			// \Mail::to($request->email_address)->send(new \App\Mail\LoginOtp($email_data));
 			return $this->sendSuccess('OTP SENT SUCCESSFULLY', ['verifaction_otp' => $verifaction_otp, 'mobile' => $request->mobile]);
 		} catch (\Throwable $e) {
@@ -91,13 +144,110 @@ class AuthController extends BaseController
 		}
 	}
 
+	public function reSentOtpMobileUpdate(Request $request)
+	{
+		$error_message = 	[
+			'mobile.required'  	=> 'Mobile address should be required',
+		];
+		$rules = [
+			'mobile' => 'required|unique:users,mobile,' . auth()->user()->id,
+		];
+		$validator = Validator::make($request->all(), $rules, $error_message);
+		if ($validator->fails()) {
+			return $this->sendFailed($validator->errors()->first(), 200);
+		}
+		$mobileExist = User::where('id', '!=', auth()->user()->id)->where(['role' => auth()->user()->role, 'mobile' => $request->mobile])->count();
+		if ($mobileExist > 0) {
+			return $this->sendFailed("Mobile number has been already taken", 200);
+		}
+		try {
+			$verifaction_otp = rand(1000, 9999);
+			Self::send_sms_otp($request->mobile, $verifaction_otp);
+			$user = auth()->user();
+			$user->otp = $verifaction_otp;
+			$user->save();
+			return $this->sendSuccess('OTP SENT SUCCESSFULLY', ['verifaction_otp' => $verifaction_otp, 'mobile' => $request->mobile]);
+		} catch (\Throwable $e) {
+			return $this->sendFailed($e->getMessage() . ' on line ' . $e->getLine(), 400);
+		}
+	}
+
+	public function registerReSentOtp(Request $request)
+	{
+		$error_message = 	[
+			// 'mobile.unique'  	              => 'mobile has been already taken',
+			'mobile.required'            	  => 'Mobile should be required',
+			'user_id.required'			 	  => 'User Id should be required',
+		];
+		$rules = [
+			'mobile'                          => 'required|min:10|max:10|exists:users,mobile',
+			'user_id'						  => 'required|integer|exists:users,id',
+		];
+		$validator = Validator::make($request->all(), $rules, $error_message);
+		if ($validator->fails()) {
+			return $this->sendFailed($validator->errors()->first(), 200);
+		}
+		$user_detail = User::find($request->user_id);
+		if ($user_detail->mobile != $request->mobile) {
+			return $this->sendFailed("Mobile number does not exist", 200);
+		}
+		try {
+			$verifaction_otp = rand(1000, 9999);
+			Self::send_sms_otp($request->mobile, $verifaction_otp);
+			$user_detail->otp = $verifaction_otp;
+			$user_detail->save();
+			Self::send_sms_otp($request->mobile, $verifaction_otp);
+			\DB::commit();
+			return $this->sendSuccess('OTP SENT SUCCESSFULLY', ['user_id' => $user_detail->id, 'otp' => $verifaction_otp]);
+		} catch (\Throwable $e) {
+			return $this->sendFailed($e->getMessage() . ' on line ' . $e->getLine(), 400);
+		}
+	}
+
+	public function updateMobile(Request $request)
+	{
+		$error_message = 	[
+			'mobile.required'  	=> 'Mobile should be required',
+			'OTP.required'  	=> 'OTP  should be required',
+		];
+		$rules = [
+			'mobile'            => 'required|min:10|max:10',
+			'otp'               => 'required|min:4|max:4',
+		];
+		$validator = Validator::make($request->all(), $rules, $error_message);
+		if ($validator->fails()) {
+			return $this->sendFailed($validator->errors()->first(), 200);
+		}
+		$mobileExist = User::where('id', '!=', auth()->user()->id)->where(['role' => auth()->user()->role, 'mobile' => $request->mobile])->count();
+		if ($mobileExist > 0) {
+			return $this->sendFailed("Mobile number has been already taken", 200);
+		}
+
+		if (auth()->user()->otp != $request->otp) {
+			return $this->sendFailed("wrong otp", 200);
+		}
+
+		try {
+			\DB::beginTransaction();
+			$user_details = auth()->user();
+			$user_details->mobile = $request->mobile;
+			$user_details->otp = null;
+			$user_details->save();
+			\DB::commit();
+			return $this->sendSuccess('MOBILE NUMBER UPDATE SUCCESSFULLY');
+		} catch (\Throwable $e) {
+			\DB::rollback();
+			return $this->sendFailed($e->getMessage() . ' on line ' . $e->getLine(), 400);
+		}
+	}
+
 	// UPDATE PROFILE
 	public function updateUserProfile(Request $request)
 	{
-		$user_details = auth()->user();
 		$error_message = 	[
 			'name.required'    	 => 'Name should be required',
 			'email.required'	 => 'Email address should be required',
+			'address.required'	 => 'Address should be required',
 			'email.unique'  	 => 'Email address has been taken',
 			'profile_pic.mimes'  => 'Profile photo format jpg,jpeg,png',
 			'profile_pic.max'    => 'Profile photo max size 2 MB',
@@ -106,21 +256,33 @@ class AuthController extends BaseController
 		];
 		$rules = [
 			'name'            => 'required|max:20',
-			'email'           => 'required|email',
+			'email' 		  => 'required|email',
+			'mobile'          => 'required|min:10|max:10',
 			'dob'			  => 'required|date',
+			'address'		  => 'required',
 		];
 		if (!empty($request->profile_pic)) {
 			$rules['profile_pic'] = 'mimes:jpg,jpeg,png|max:2000';
 		}
 		$validator = Validator::make($request->all(), $rules, $error_message);
 		if ($validator->fails()) {
-			return $this->sendFailed(implode(", ", $validator->errors()->all()), 200);
+			return $this->sendFailed($validator->errors()->first(), 200);
 		}
-		$emailExist = User::where('id', '!=', auth()->user()->id)->where(['email' => $request->email])->count();
+
+		$mobileExist = User::where('id', '!=', auth()->user()->id)->where(['role' => auth()->user()->role, 'mobile' => $request->mobile])->count();
+		if ($mobileExist > 0) {
+			return $this->sendFailed("Mobile number has been already taken", 200);
+		}
+
+		$emailExist = User::where('id', '!=', auth()->user()->id)->where(['role' => auth()->user()->role, 'email' => $request->email])->count();
 		if ($emailExist > 0) {
 			return $this->sendFailed("Email address has been already taken", 200);
 		}
+
 		try {
+			\DB::beginTransaction();
+			$user_details = auth()->user();
+			$user_details->fill($request->only('name', 'email', 'address', 'dob', 'address'));
 			if (!empty($request->file('profile_pic'))) {
 				if (Storage::disk('public')->exists('user_images/' . $user_details->user_pic_name)) {
 					Storage::disk('public')->delete('user_images/' . $user_details->user_pic_name);
@@ -128,21 +290,28 @@ class AuthController extends BaseController
 				$user_pic = time() . '_' . rand(1111, 9999) . '.' . $request->file('profile_pic')->getClientOriginalExtension();
 				$request->file('profile_pic')->storeAs('user_images', $user_pic, 'public');
 				$request['profile_pic'] = $user_pic;
+				$user_details->profile_pic = $user_pic;
 			}
-
-			\DB::beginTransaction();
-			$user_details->fill($request->all());
-			$user_details->profile_pic = $user_pic;
 			$user_details->save();
 			\DB::commit();
-			return $this->sendSuccess('PROFILE UPDATED SUCCESSFULLY', new UserProfileCollection($user_details));
+			if ($request->mobile == auth()->user()->mobile) {
+				$mobileVerify = array('mobile_verify_status' => 1);
+				return $this->sendSuccess('PROFILE UPDATE SUCCESSFULLY', $mobileVerify);
+			} else {
+				$otp = rand(1000, 9999);
+				Self::send_sms_otp($request->mobile, $otp);
+				$user = auth()->user();
+				$verifaction_otp = $otp;
+				$user->otp = $verifaction_otp;
+				$user->save();
+				$mobileVerify = array('mobile_verify_status' => 0, 'otp' => $otp);
+				return $this->sendSuccess('PROFILE UPDATE AND OTP SENT SUCCESSFULLY YOUR MOBILE', $mobileVerify);
+			}
 		} catch (\Throwable $e) {
 			\DB::rollback();
 			return $this->sendFailed($e->getMessage() . ' on line ' . $e->getLine(), 400);
 		}
 	}
-
-
 
 	public function forgot_password(Request $request)
 	{
@@ -162,7 +331,7 @@ class AuthController extends BaseController
 			if (!isset($user_detail)) {
 				return $this->sendFailed('WE COULD NOT FOUND ANY ACCOUNT', 200);
 			}
-			$verifaction_otp = rand(111111, 999999);
+			$verifaction_otp = rand(1000, 9999);
 			$email_data = ['user_name' => $user_detail->first_name, 'verifaction_otp' => $verifaction_otp];
 			\Mail::to($user_detail->email)->send(new \App\Mail\ForgotPassword($email_data));
 			return $this->sendSuccess('OTP SENT SUCCESSFULLY', ['user_id' => $user_detail->id, 'verifaction_otp' => $verifaction_otp, 'email' => $user_detail->email]);
@@ -195,6 +364,33 @@ class AuthController extends BaseController
 			$user_detail->save();
 			\DB::commit();
 			return $this->sendSuccess('PASSWORD UPDATED SUCCESSFULLY');
+		} catch (\Throwable $e) {
+			\DB::rollback();
+			return $this->sendFailed($e->getMessage() . ' on line ' . $e->getLine(), 400);
+		}
+	}
+
+	public function send_sms_otp($mobile_number, $verification_otp)
+	{
+
+		// $opt_url = "https://2factor.in/API/V1/fd9c6a99-19d7-11ec-a13b-0200cd936042/SMS/" . $mobile_number . "/" . $verification_otp . "/OTP_TAMPLATE";
+		// $curl = curl_init();
+		// curl_setopt($curl, CURLOPT_URL, $opt_url);
+		// curl_setopt($curl, CURLOPT_RETURNTRANSFER, TRUE);
+		// curl_setopt($curl, CURLOPT_PROXYPORT, "80");
+		// curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, false);
+		// $result = curl_exec($curl);
+		return;
+	}
+
+
+	public function changeOnlineStatus()
+	{
+		try {
+			\DB::beginTransaction();
+			$change = User::find(auth()->user()->id)->update(['online_status' => auth()->user()->online_status == 'Online' ? 'Offline' : 'Online']);
+			\DB::commit();
+			return $this->sendSuccess('Online Status change succssfully');
 		} catch (\Throwable $e) {
 			\DB::rollback();
 			return $this->sendFailed($e->getMessage() . ' on line ' . $e->getLine(), 400);
